feature(home): added map re-centering when selecting a report
